adjust friction to force ratio in the physics properties.
