--- HamsterWheel/Drag and Drop Game/Nodes/Match1.swift
@@ -1,35 +1,0 @@
-//
-//  Match.swift
-//  HamsterWheel
-//
-//  Created by Phyllis Wong on 3/6/18.
-//  Copyright © 2018 Bob De Kort. All rights reserved.
-//
-
-import Foundation
-import SpriteKit
-
-class Match1: SKSpriteNode {
-    
-    init(imageNamed: String) {
-        let size = CGSize(width: 150, height: 150)
-        let texture = SKTexture(imageNamed: imageNamed)
-        
-        super.init(texture: texture, color: .white, size: size)
-        setupMatch1Physics()
-    }
-    
-    func setupMatch1Physics() {
-        
-        self.physicsBody = SKPhysicsBody(circleOfRadius: 1)
-        self.physicsBody?.affectedByGravity = false
-        self.physicsBody?.categoryBitMask = PhysicsCategory.Match1
-        self.physicsBody?.collisionBitMask = PhysicsCategory.Wall
-        self.physicsBody?.contactTestBitMask = PhysicsCategory.Player1
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        
-        fatalError("init(coder:) has not been implemented")
-    }
-}
