+++ HamsterWheel/Drag and Drop Game/Nodes/Player.swift
@@ -23,6 +23,7 @@ class Player1: SKSpriteNode {
         
         self.physicsBody = SKPhysicsBody(circleOfRadius: 50)
         self.physicsBody?.affectedByGravity = false
+        self.physicsBody?.linearDamping = 1
         self.physicsBody?.categoryBitMask = PhysicsCategory.Player1
         self.physicsBody?.collisionBitMask = PhysicsCategory.Wall
         self.physicsBody?.contactTestBitMask = PhysicsCategory.Match1
@@ -32,3 +33,28 @@ class Player1: SKSpriteNode {
         fatalError("init(coder:) has not been implemented")
     }
 }
+
+class Player2: SKSpriteNode {
+    
+    init(imageNamed: String) {
+        let size = CGSize(width: 100, height: 100)
+        let texture = SKTexture(imageNamed: imageNamed)
+        super.init(texture: texture, color: .white, size: size)
+        setupPlayer2Physics()
+    }
+    
+    func setupPlayer2Physics() {
+        
+        self.physicsBody = SKPhysicsBody(circleOfRadius: 50)
+        self.physicsBody?.affectedByGravity = false
+        self.physicsBody?.linearDamping = 1
+        self.physicsBody?.categoryBitMask = PhysicsCategory.Player2
+        self.physicsBody?.collisionBitMask = PhysicsCategory.Wall
+        self.physicsBody?.contactTestBitMask = PhysicsCategory.Match2
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
